rename and reshuffle some hooks
