+++ client/machines/timerSettings/model.ts
@@ -2,7 +2,7 @@ import { UserConfig } from '@shared/types';
 import { ContextFrom, EventFrom } from 'xstate';
 import { createModel } from 'xstate/lib/model';
 
-const timerSettingsModel = createModel(
+export const timerSettingsModel = createModel(
   {
     pomo: 1,
     short: 1,
@@ -20,5 +20,3 @@ const timerSettingsModel = createModel(
 
 export type TimerSettingsContext = ContextFrom<typeof timerSettingsModel>;
 export type TimerSettingsEvents = EventFrom<typeof timerSettingsModel>;
-
-export default timerSettingsModel;
